BG/fixed header bugs and renamed without known component
