+++ assets/css/font.css
@@ -120,30 +120,35 @@
 }
 .fa-rotate-90 {
   -ms-filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=1)';
+  filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=1)';
   -webkit-transform: rotate(90deg);
   -ms-transform: rotate(90deg);
   transform: rotate(90deg);
 }
 .fa-rotate-180 {
   -ms-filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=2)';
+  filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=2)';
   -webkit-transform: rotate(180deg);
   -ms-transform: rotate(180deg);
   transform: rotate(180deg);
 }
 .fa-rotate-270 {
   -ms-filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=3)';
+  filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=3)';
   -webkit-transform: rotate(270deg);
   -ms-transform: rotate(270deg);
   transform: rotate(270deg);
 }
 .fa-flip-horizontal {
   -ms-filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=0, mirror=1)';
+  filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=0, mirror=1)';
   -webkit-transform: scale(-1, 1);
   -ms-transform: scale(-1, 1);
   transform: scale(-1, 1);
 }
 .fa-flip-vertical {
   -ms-filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=2, mirror=1)';
+  filter: 'progid:DXImageTransform.Microsoft.BasicImage(rotation=2, mirror=1)';
   -webkit-transform: scale(1, -1);
   -ms-transform: scale(1, -1);
   transform: scale(1, -1);
