+++ components/landing/skills.vue
@@ -65,3 +65,9 @@
 		</section>
   </div>
 </template>
+
+<script>
+export default {
+	name: 'skills'
+}
+</script>
